Refactor transfer stat handler to its own class
